split whitelabel examples by endpoint
